Amplia_columnas

Description

Las matrices enteras se pueden representar mediante tablas con índices enteros:

  type Matriz = Array (Int,Int) Int

Definir la función

  ampliaColumnas :: Matriz -> Matriz -> Matriz

tal que (ampliaColumnas p q) es la matriz construida añadiendo las columnas de la matriz q a continuación de las de p (se supone que tienen el mismo número de filas). Por ejemplo, si p y q representan las dos primeras matrices, entonces (ampliaColumnas p q) es la tercera

  |0 1|    |4 5 6|    |0 1 4 5 6| 
  |2 3|    |7 8 9|    |2 3 7 8 9|

En Haskell,

>>> let p = listArray ((1,1),(2,2)) [0..3]
>>> let q = listArray ((1,1),(2,3)) [4..9]
>>> elems (ampliaColumnas p q)
[0,1,4,5,6,2,3,7,8,9]
>>> bounds (ampliaColumnas p q)
((1,1),(2,5))

Documentation

type Matriz = Array (Int, Int) Int

Tipo de las matrices.

ampliaColumnas :: Matriz -> Matriz -> Matriz

Definición.