Divisores_con_final

Description

Divisores de un número con final dado

Definir la función

divisoresConFinal :: Integer -> Integer -> [Integer]

tal que (divisoresConFinal n m) es la lista de los divisores de n cuyos dígitos finales coincide con m. Por ejemplo,

>>> divisoresConFinal 84 4
[4,14,84]
>>> divisoresConFinal 720 20
[20,120,720]

Documentation

divisoresConFinal :: Integer -> Integer -> [Integer]

1ª definición.

final :: Integer -> Integer -> Bool

(final x y) se verifica si las cifras finales de x coincide con y. Por ejemplo,

>>> final 325 5
True
>>> final 325 25
True
>>> final 325 35
False

divisoresConFinal2 :: Integer -> Integer -> [Integer]

2ª solución (con isSuffixOf).

prop_equiv_divisoresConFinal :: Positive Integer -> Positive Integer -> Bool

Comprobación de la equivalencia de las definiciones de divisoresConFinal.

>>> quickCheck prop_equiv_divisoresConFinal
+++ OK, passed 100 tests.