ElementosMinimales

Description

Definir la función

minimales :: Eq a => [[a]] -> [[a]]

tal que (minimales xss) es la lista de los elementos de xss que no están contenidos en otros elementos de xss. Por ejemplo,

>>> minimales [[1,3],[2,3,1],[3,2,5]]
[[2,3,1],[3,2,5]]
>>> minimales [[1,3],[2,3,1],[3,2,5],[3,1]]
[[2,3,1],[3,2,5]]

Documentation

minimales :: Eq a => [[a]] -> [[a]]

Definición

subconjuntoPropio :: Eq a => [a] -> [a] -> Bool

(subconjuntoPropio xs ys) se verifica si xs es un subconjunto propio de ys. Por ejemplo,

>>> subconjuntoPropio [1,3] [3,1,3]
False
>>> subconjuntoPropio [1,3,1] [3,1,2]
True