Emparejamiento_binario

Description

Definir la función

zipBinario :: [a -> b -> c] -> [a] -> [b] -> [c]

tal que (zipBinario fs xs ys) es la lista obtenida aplicando cada una de las operaciones binarias de fs a los correspondientes elementos de xs e ys. Por ejemplo,

>>> zipBinario [(+), (*), (^)] [2,2,2] [4,4,4]
[6,8,16]
>>> zipBinario [(+)] [2,2,2] [4,4,4]
[6]
>>> zipBinario (cycle [(+), (*)]) [1 .. 4] [2..5]
[3,6,7,20]

Documentation

zipBinario :: [a -> b -> c] -> [a] -> [b] -> [c]

1ª definición (por recursión).

zipBinario2 :: [a -> b -> c] -> [a] -> [b] -> [c]

2ª definición (con zip).

zipBinario3 :: [a -> b -> c] -> [a] -> [b] -> [c]

3ª definición (con zip3).

zipBinario4 :: [a -> b -> c] -> [a] -> [b] -> [c]

4ª definición (con zipWith3).

newtype Operacion

Tipo de operaciones para la verificación.

Constructors

O (Int -> Int -> Int)

gen_operaciones :: Gen [Operacion]

Generador de operaciones.

prop_zipBinario :: [Int] -> [Int] -> Property

Comprueba la equivalencia de las definiciones de zipBinario.

>>> quickCheck prop_zipBinario
+++ OK, passed 100 tests.