Mas_repetido

Description

Definir la función

masRepetido :: Ord a => [a] -> (a,Int)

tal que (masRepetido xs) es el mayor elemento de xs que aparece más veces de manera consecutiva en la lista junto con el número de sus apariciones consecutivas. Por ejemplo,

>>> masRepetido [1,1,4,4,1]
(4,2)
>>> masRepetido [4,4,1,1,4]
(4,2)
>>> masRepetido "aadda"
('d',2)
>>> masRepetido "a"
('a',1)
>>> masRepetido "ba"
('b',1)

Documentation

masRepetido :: Ord a => [a] -> (a, Int)

1ª definición (por recursión)

masRepetidos :: Ord a => [a] -> [(a, Int)]

masRepetido2 :: Ord a => [a] -> (a, Int)

2ª definición (con group y maximum).

masRepetido3 :: Ord a => [a] -> (a, Int)

3ª definición (con group, maximum y swap).

masRepetido4 :: Ord a => [a] -> (a, Int)

4ª definición (con group y maximumBy).

masRepetido5 :: Ord a => [a] -> (a, Int)

5ª definición (sin argumentos)

prop_masRepetido :: NonEmptyList Int -> Bool

(prop_masRepetido xs) se verifica si las definiciones de masRepetido son equivalentes sobre xs. Por ejemplo,

>>> quickCheck prop_masRepetido
+++ OK, passed 100 tests.