Mayor_producto_de_las_ramas_de_un_arbol

Description

Los árboles se pueden representar mediante el siguiente tipo de datos

   data Arbol a = N a [Arbol a]
                  deriving Show

Por ejemplo, los árboles

      1              3
    /  \            /|\ 
   2   3           / | \
       |          5  4  7
       4          |     /\ 
                  6    2  1

se representan por

  ej1, ej2 :: Arbol Int
  ej1 = N 1 [N 2 [],N 3 [N 4 []]]
  ej2 = N 3 [N 5 [N 6 []], N 4 [], N 7 [N 2 [], N 1 []]]

Definir la función

mayorProducto :: (Ord a, Num a) => Arbol a -> a

tal que (mayorProducto a) es el mayor producto de las ramas del árbol a. Por ejemplo,

>>> mayorProducto (N 1 [N 2 [], N  3 []])
3
>>> mayorProducto (N 1 [N 8 [], N  4 [N 3 []]])
12
>>> mayorProducto (N 1 [N 2 [],N 3 [N 4 []]])
12
>>> mayorProducto (N 3 [N 5 [N 6 []], N 4 [], N 7 [N 2 [], N 1 []]])
90
>>> mayorProducto (N (-8) [N 0 [N (-9) []],N 6 []])
0
>>> let a = N (-4) [N (-7) [],N 14 [N 19 []],N (-1) [N (-6) [],N 21 []],N (-4) []]
>>> mayorProducto a
84

Documentation

data Arbol a

Tipo de árboles

Constructors

N a [Arbol a]

mayorProducto :: (Ord a, Num a) => Arbol a -> a

1º definición.

productosRamas :: (Ord a, Num a) => Arbol a -> [a]

(productosRamas a) es la lista de los productos de las ramas del árbol a. Por ejemplo,

>>> productosRamas (N 3 [N 5 [N 6 []], N 4 [], N 7 [N 2 [], N 1 []]])
[90,12,42,21]

mayorProducto2 :: (Ord a, Num a) => Arbol a -> a

2ª definición.

menorProducto :: (Ord a, Num a) => Arbol a -> a

(menorProducto a) es el menor producto de las ramas del árbol a. Por ejemplo,

>>> menorProducto (N 1 [N 2 [], N  3 []])
2
>>> menorProducto (N 1 [N 8 [], N  4 [N 3 []]])
8
>>> menorProducto (N 1 [N 2 [],N 3 [N 4 []]])
2
>>> menorProducto (N 3 [N 5 [N 6 []], N 4 [], N 7 [N 2 [], N 1 []]])
12

mayorProducto3 :: (Ord a, Num a) => Arbol a -> a

3ª definición.

ramas :: Arbol b -> [[b]]

(ramas a) es la lista de las ramas del árbol a. Por ejemplo,

>>> ramas (N 3 [N 5 [N 6 []], N 4 [], N 7 [N 2 [], N 1 []]])
[[3,5,6],[3,4],[3,7,2],[3,7,1]]

mayorProducto4 :: (Ord a, Num a) => Arbol a -> a

4ª definición.

mayorProducto5 :: (Ord a, Num a) => Arbol a -> a

5ª definición.

gen_Arbol :: Arbitrary a => Int -> Gen (Arbol a)

Generador de árboles.

> sample ((gen_Arbol 5) :: Gen (Arbol Int))
N 0 [N 0 []]
N (-2) []
N 4 []
N 2 [N 4 []]
N 8 []
N (-2) [N (-9) [],N 7 []]
N 11 []
N (-11) [N 4 [],N 14 []]
N 10 [N (-3) [],N 13 []]
N 12 [N 11 []]
N 20 [N (-18) [],N (-13) []]

prop_mayorProducto :: Arbol Int -> Bool

Comprueba la equivalencia de las definiciones de mayorProducto.

>>> quickCheck prop_mayorProducto
+++ OK, passed 100 tests.