PimPamPum

Description

Pim, Pam, Pum y divisibilidad

Definir la función

sonido :: Int -> String

tal que (sonido n) escribe Pim si n es divisible por 3, además escribe Pam si n es divisible por 5 y también escribe Pum si n es divisible por 7. Por ejemplo,

>>> sonido   3
"Pim"
>>> sonido   5
"Pam"
>>> sonido   7
"Pum"
>>> sonido   8
""
>>> sonido   9
"Pim"
>>> sonido  15
"PimPam"
>>> sonido  21
"PimPum"
>>> sonido  35
"PamPum"
>>> sonido 105
"PimPamPum"

Documentation

sonido :: Int -> String

1ª solución.

sonido2 :: Int -> String

2ª solución.

sonido3 :: Int -> String

3ª solución.

sonido4 :: Int -> String

4ª solución.

prop_equiv_sonido :: Int -> Bool

(prop_equiv_sonido x) se verifica si las definiciones de sonido son equivalentes sobre x. Por ejemplo,

>>> all prop_equiv_sonido [3,5,7,9,15,21,35,105]
True

verifica_equiv_sonido :: IO ()

Comprueba la equivalencia de las definiciones de sonido.

>>> verifica_equiv_sonido
+++ OK, passed 100 tests.