Primos_consecutivos_con_media_capicua

Description

Definir las funciones

primosConsecutivosConMediaCapicua :: [(Int,Int,Int)]
nPrimosConsecutivosConMediaCapicua :: Int -> Int

tales que

primosConsecutivosConMediaCapicua es la lista de las ternas (x,y,z) tales que x e y son primos consecutivos cuya media, z, es capicúa. Por ejemplo,

>>> take 5 primosConsecutivosConMediaCapicua
[(3,5,4),(5,7,6),(7,11,9),(97,101,99),(109,113,111)]

(nPrimosConsecutivosConMediaCapicua n) es el número de ternas de primos consecutivos con media capicua que son menores que n. Por ejemplo,

>>> nPrimosConsecutivosConMediaCapicua 2014
20

Documentation

primosConsecutivosConMediaCapicua :: [(Int, Int, Int)]

1ª definición (definiendo los primos)

primo :: Int -> Bool

(primo x) se verifica si x es primo. Por ejemplo, >>> primo 7 True >>> primo 8 False

primos :: [Int]

primos es la lista de los números primos mayores que 2. Por ejemplo,

>>> take 10 primos
[3,5,7,11,13,17,19,23,29,31]

capicua :: Int -> Bool

(capicua x) se verifica si x es capicúa. Por ejemplo,

>>> capicua 32723
True
>>> capicua 32732
False

primosConsecutivosConMediaCapicua2 :: [(Int, Int, Int)]

2ª definición (con primes)

verifica_primosConsecutivosConMediaCapicua :: Int -> Bool

(verifica_primosConsecutivosConMediaCapicua n =) se verifica si las definiciones de primosConsecutivosConMediaCapicua son equivalentes para los n primeros elementos. Por ejemplo,

>>> verifica_primosConsecutivosConMediaCapicua 20
True

nPrimosConsecutivosConMediaCapicua :: Int -> Int

Definición de nPrimosConsecutivosConMediaCapicua