Regiones

Description

En los siguientes dibujos se observa que el número máximo de regiones en el plano generadas con 1, 2 ó 3 líneas son 2, 4 ó 7, respectivamente.

                     \  |
                      \5|
                       \|
                        \
                        |\
                        | \
              |         |  \ 
   1        1 | 3     1 | 3 \  6
  ------   ---|---   ---|----\---
   2        2 | 4     2 | 4   \ 7
              |         |      \

Definir la función

regiones :: Integer -> Integer  

tal que (regiones n) es el número máximo de regiones en el plano generadas con n líneas. Por ejemplo,

>>> regiones 3
7  
>>> regiones 100
5051

Documentation

regiones :: Integer -> Integer

1ª definición (por recursión).

regiones2 :: Integer -> Integer

2ª definición (por la fórmula).

prop_regiones :: Positive Integer -> Bool

(prop_regiones n) se verifica si las definiciones de regiones son equivalentes sobre n. Por ejemplo,

>>> all prop_regiones [Positive 2, Positive 100]
True
>>> quickCheck prop_regiones
+++ OK, passed 100 tests.