Trenza

Description

Trenzado de listas

Definir la función

trenza :: [a] -> [a] -> [a]

tal que (trenza xs ys) es la lista obtenida intercalando los elementos de xs e ys. Por ejemplo,

>>> trenza [5,1] [2,7,4]
[5,2,1,7]
>>> trenza [5,1,7] [2..]
[5,2,1,3,7,4]
>>> trenza [2..] [5,1,7]
[2,5,3,1,4,7]
>>> take 8 (trenza [2,4..] [1,5..])
[2,1,4,5,6,9,8,13]

Documentation

trenza :: [a] -> [a] -> [a]

1ª definición (por comprensión).

trenza2 :: [a] -> [a] -> [a]

2ª definición (con zipWith).

trenza3 :: [a] -> [a] -> [a]

3ª definición (con zipWith y sin argumentos).

trenza4 :: [a] -> [a] -> [a]

4ª definición (por recursión).

prop_trenza :: Eq a => [a] -> [a] -> Bool

prop_trenza xs ys) se verifica si las definiciones de trenza son equivalentes para xs e ys. Por ejemplo,

>>> :{
 and [ prop_trenza [5,1] [2,7,4]
     , prop_trenza [5,1,7] [2..]
     , prop_trenza [2..] [5,1,7] ]
:}
True

verifica_trenza :: IO ()

Comprueba la equivalencia de las definiciones de trenza.

>>> verifica_trenza
+++ OK, passed 100 tests.